+++ tools/configure_esp_secure_cert.py
@@ -302,11 +302,13 @@ def main():
         if args.esp_secure_cert_csv is not None:
             esp_secure_cert.parse_esp_secure_cert_csv(args.esp_secure_cert_csv)
 
-        bin_filename = esp_secure_cert.generate_esp_secure_cert(args.target_chip, args.port)
 
         if args.secure_sign:
+            bin_filename = esp_secure_cert.generate_esp_secure_cert(args.target_chip, args.port, add_tlv_integrity=False)
             # Set the secure boot scheme
             bin_filename = esp_secure_cert.add_signature_block_using_existing_key(bin_filename, args.signing_key_file, args.signing_scheme)
+
+        bin_filename = esp_secure_cert.generate_esp_secure_cert(args.target_chip, args.port, add_tlv_integrity=True)
 
         if not args.skip_flash:
             esp_secure_cert.flash_esp_secure_cert_partition(args.target_chip, args.port, args.sec_cert_part_offset, bin_filename)